Redid client config command.
To avoid sending someone elses config from one player to another, all client configs have been move to their own config file and class. It now sends a packet with the value that's being changed to client, where it is updated and save since client side configs don't need to be synced with the server.